Substitution in shorthands:     when evaluating the expression in <>,     we need to have access to the local variables at the point     of the shorthand call (n(), tempo(), etc.).     Pass the locals explicitly (as locals()), and pass these to eval().     Maybe there's some better way of doing this; I can't find one.
Change the meaning of Note.nchord and Note.chord_pos
Previously they were relative to all the notes starting at this time.
But in many cases this will be a combination of chord and melody notes.
Chord voicing should not apply to the melody notes.
So have nchord and chord_pos be relative only to notes tagged "ch",
and have [ ... ] in score notation add this tag.