add IPA_LAYER type
controls a nuance layer. values
'on': show vars and do the nuance
'hide': hide vars but do the nuance
'off': hide vars and don't do the nuance
Use it as a tag in variables specific to that layer,
and use it to condition doing the nuance, e.g.
if foo != 'off':
    ns.vol_adjust_pft(...)